refactor: 불필요한 ShortestPathFinder.addEdge() 인터페이스 제거
