+++ src/main/java/nextstep/subway/path/domain/PathFinder.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.Objects;
 
 import nextstep.subway.line.domain.Line;
-import nextstep.subway.line.domain.Section;
 import nextstep.subway.path.domain.shortest.DijkstraShortestPathFinder;
 import nextstep.subway.path.domain.shortest.ShortestPath;
 import nextstep.subway.path.domain.shortest.ShortestPathFinder;
@@ -15,26 +14,7 @@ public class PathFinder {
 	private final ShortestPathFinder shortestPathFinder;
 
 	private PathFinder(List<Line> lines) {
-		validate(lines);
-
-		this.shortestPathFinder = DijkstraShortestPathFinder.of();
-		lines.stream()
-			.flatMap(line -> line.getSections().stream())
-			.forEach(this::addSection);
-	}
-
-	private void validate(List<Line> lines) {
-		if (null == lines) {
-			throw new IllegalArgumentException("노선 목록이 있어야 합니다.");
-		}
-	}
-
-	private void addSection(Section section) {
-		shortestPathFinder.addEdge(
-			section.getUpStation(),
-			section.getDownStation(),
-			section.getDistance()
-		);
+		this.shortestPathFinder = DijkstraShortestPathFinder.of(lines);
 	}
 
 	public ShortestPath findShortest(Station source, Station target) {
